Set max_age on Enterprise cookie.
We set a cookie which identifies a user as being associated with
a given enterprise based on how they navigated to the Open edX site.
This cookie gets shared across edX services, so that the enterprise
context can be maintained when navigating between UIs across these
services. This cookie needs to be short-lived so that the Enterprise
context does not linger longer than it should. This change sets a
60 second max age on the Enterprise cookie.

ENT-367